ai-knowledge-agent/requirements.txt
streamlit
google-generativeai
numpy
pillow
PyPDF2
pymupdf
python-dotenv
langchain-text-splitters
faiss-cpu
pytesseract
